\name{elife_paper}
\alias{elife_paper}
\title{Get full text eLife papers via XML.}
\usage{
  elife_paper(doi = NULL)
}
\arguments{
  \item{doi}{DOI to get full text for.}
}
\description{
  Get full text eLife papers via XML.
}
\details{
  Returns raw XML right now.
}
\examples{
\dontrun{
elife_paper(doi="10.7554/eLife.00160")
}
}

